Improve testing
- add pytest-cov
- improve test coverage
