+++ src/gpmicrobiome/__init__.py
@@ -1,4 +1,5 @@
 """__init__.py."""
-from gpmicrobiome.src import get_mcmc_summary, run
+from gpmicrobiome.inference import run_nuts
+from gpmicrobiome.utils import get_mcmc_summary
 
-__all__ = ["run", "get_mcmc_summary"]
+__all__ = ["run_nuts", "get_mcmc_summary"]
