+++ src/gpmicrobiome/constants.py
@@ -1,0 +1,4 @@
+"""constants.py."""
+
+GP_INPUT_DIMENSIONS = 2
+OTU_MATRIX_DIMENSIONS = 2
