+++ src/gpmicrobiome/models.py
@@ -1,23 +1,13 @@
-"""src.py."""
-from typing import Any, Mapping
-
+"""models.py."""
 import jax.numpy as jnp
-import numpy as np
-import numpy.typing as npt
 import numpyro
 import numpyro.distributions as dist
-import pandas as pd
-from jax import Array, random
+from jax import Array
 from jax.nn import softmax
-from numpyro.diagnostics import summary
-from numpyro.infer import MCMC, NUTS
 
-# ruff: noqa: N803, N806, F841, PLR0913, PLR0917
+from gpmicrobiome.constants import GP_INPUT_DIMENSIONS
 
-KeyArray = Array
-
-GP_INPUT_DIMENSIONS = 2
-OTU_MATRIX_DIMENSIONS = 2
+# ruff: noqa: N803, N806, F841, PLR0914
 
 
 def get_default_priors() -> dict[str, dist.Distribution]:
@@ -199,10 +189,9 @@ def model(
         if use_deterministic:
             theta_n = numpyro.deterministic("theta_n", theta_n)
 
-    with numpyro.plate("otu", y.shape[-1]), numpyro.plate("time", y.shape[-2]):
-        b = numpyro.sample("b", priors["b"])
-
     if consider_zero_inflation:
+        with numpyro.plate("otu", y.shape[-1]), numpyro.plate("time", y.shape[-2]):
+            b = numpyro.sample("b", priors["b"])
         theta_b = (tmp := theta_n * b) / jnp.sum(tmp, axis=-1, keepdims=True)
         if use_deterministic:
             theta_b = numpyro.deterministic("theta_b", theta_b)
@@ -218,119 +207,3 @@ def model(
         ),
         obs=y,
     )
-
-
-def get_mcmc_summary(mcmc: MCMC) -> pd.DataFrame:
-    """Get MCMC summary DataFrame.
-
-    Args:
-        mcmc: MCMC object.
-
-    Returns:
-        DataFrame.
-    """
-
-    def process_variable(
-        variable: str, data: Mapping[str, npt.NDArray[np.float64] | np.float64]
-    ) -> dict[str, Any]:
-        res: dict[str, str | list[tuple[int, ...]] | list[np.float64] | list[None]] = {
-            "variable": variable
-        }
-        for statistic, values in data.items():
-            if "index" not in res:
-                if isinstance(values, np.ndarray):
-                    res["index"] = [
-                        tuple(map(int, x))
-                        for x in zip(
-                            *jnp.unravel_index(jnp.arange(values.size), values.shape)
-                        )
-                    ]
-                else:
-                    res["index"] = [None]
-            if isinstance(values, np.ndarray):
-                res[statistic] = values.flatten().tolist()
-            else:
-                res[statistic] = [values]
-        return res
-
-    return pd.concat(
-        [
-            pd.DataFrame.from_dict(process_variable(variable, data))
-            for variable, data in summary(mcmc.get_samples(group_by_chain=True)).items()
-        ],
-        axis=0,
-        ignore_index=True,
-    )
-
-
-def run(
-    key: KeyArray,
-    X: Array,
-    y: Array,
-    X_pred: Array | None = None,
-    reference_category: int | Array | None = None,
-    priors: dict[str, dist.Distribution] | None = None,
-    num_basis: int = 10,
-    num_warmup: int = 1_000,
-    num_samples: int = 1_000,
-    num_chains: int = 4,
-    use_deterministic: bool = False,
-    consider_zero_inflation: bool = True,
-) -> MCMC:
-    """Run GPMicrobiome.
-
-    HMC with NUTS.
-
-    Args:
-        key: PRNGKey.
-        X: Measurement time points (n_timepoints by 1).
-        y: OTU counts (n_timepoints by n_otus).
-        X_pred: Prediction time points (n_prediction_timepoints by 1).
-        reference_category: Reference category to make the softmax transformation bijective.
-            If not provided, then we take the middle OTU with respect to counts, i.e.
-            `reference_category = jnp.argsort(jnp.sum(y, 0))[y.shape[1] // 2]`.
-        priors: Priors distributions. Defaults to gpmicrobiome.src.get_default_priors().
-        num_basis: Number of basis functions. Defaults to 10.
-        num_warmup: Number of warmup iterations. Defaults to 1_000.
-        num_samples: Number of sampling iterations. Defaults to 1_000.
-        num_chains: Number of chains. Defaults to 4.
-        use_deterministic: Whether to include the deterministic sites in trace. Defaults to False.
-        consider_zero_inflation: Whether to consider zero inflation. Defaults to True.
-
-    Returns:
-        NumPyro MCMC object.
-    """
-    assert len(y.shape) == OTU_MATRIX_DIMENSIONS, "y should have two dimensions"
-    assert len(X.shape) == GP_INPUT_DIMENSIONS, "X should have two dimensions"
-    if X_pred is not None:
-        assert (
-            len(X_pred.shape) == GP_INPUT_DIMENSIONS
-        ), "X_pred should have two dimensions when provided"
-
-    reference_category = reference_category or (
-        jnp.argsort(jnp.sum(y, 0))[y.shape[1] // 2]
-    ).astype(int)
-
-    priors = get_default_priors() | (priors if priors is not None else {})
-
-    key, key_ = random.split(key, 2)
-    nuts_kernel = NUTS(model)
-    mcmc = MCMC(
-        nuts_kernel,
-        num_warmup=num_warmup,
-        num_samples=num_samples,
-        num_chains=num_chains,
-    )
-    key, key_ = random.split(key, 2)
-    mcmc.run(
-        key_,
-        X=X,
-        y=y,
-        num_basis=num_basis,
-        priors=priors,
-        reference_category=reference_category,
-        X_pred=X_pred,
-        use_deterministic=use_deterministic,
-        consider_zero_inflation=consider_zero_inflation,
-    )
-    return mcmc
