+++ tox.ini
@@ -1,10 +1,12 @@
 [tox]
 minversion = 4.0
 envlist =
+    clean
     py310
     py311
     py312
     lint
+    report
 skip_missing_interpreters = true
 
 [gh-actions]
@@ -13,14 +15,16 @@ python =
     3.11: py311
     3.12: py312
 
-[testenv:.pkg]
-
 [testenv]
 deps =
     pytest
+    pytest-cov
 commands =
     pip freeze
     python -m pytest {posargs}
+depends =
+    py310,py311,py312: clean
+    report: py310,py311,py312
 
 [testenv:lint]
 skip_install = true
@@ -28,3 +32,15 @@ deps =
     pre-commit
 commands =
     pre-commit run --all-files
+
+[testenv:report]
+skip_install = true
+deps = pytest-cov
+commands =
+    coverage html
+    coverage report
+
+[testenv:clean]
+deps = pytest-cov
+skip_install = true
+commands = coverage erase
